backend/cloud_functions/archive_game/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-events>=0.7.0
